fix issues in initializing models. rerun examples.ipynb
