+++ alpsqutip/operators/states/meanfield/__init__.py
@@ -4,10 +4,12 @@ from .projections import (
     project_operator_to_m_body,
     project_to_n_body_operator,
 )
+from .variational import variational_quadratic_mfa
 
 ALL = [
     "one_body_from_qutip_operator",
     "project_meanfield",
     "project_operator_to_m_body",
     "project_to_n_body_operator",
+    "variational_quadratic_mfa",
 ]
